Transfered to github
This code has been originally hosted on:
http://source.ymc.ch/svn/ymc_datatypes/

Since this subversion-host does no longer exists, but the code might still be
usefull in the context of an article on share.ez.no, we decided to publish it
again on github.com/ymc/ymc-ezp-datatypes.

Note that this code has been unmaintained for quite some time and therefore
should not be used in a production system.